refactor(providers): move context window validation into lib/contextWindow
- Extract isCommittableWindowText + parseWindowText into lib/contextWindow.ts,
  shared by CustomProviderDialog and AddProviderWizard (removes the wizard's
  dependency on the large dialog module; Copilot suppressed suggestion)
- Dialog save path now parses via parseWindowText, keeping one source of truth
- Align manualModelWindowInvalid wording with the existing
  contextWindowInvalid error text in all four locales
- PR description updated to match final behavior: invalid non-empty input
  blocks the add with a toast and keeps the input for correction
